backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic[email]==2.6.3
email-validator==2.1.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
stripe==8.5.0
httpx